# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "此 API 文件提供社區通知系統的後端介面說明，包含使用者管理、設施預約、社區申請及訊息通知等功能。使用 JWT Bearer 進行認證。",
        "title": "社區通知系統 API 文件 (Community Notification System)",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "依登入使用者取得通知列表，支援分頁與已讀篩選。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "取得通知列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "頁碼（預設 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每頁筆數（預設 20，最大 100）",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "已讀篩選",
                        "name": "is_read",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功取得通知列表",
                        "schema": {
                            "$ref": "#/definitions/message_model.MessageListResponse"
                        }
                    },
                    "400": {
                        "description": "請求參數錯誤",
                        "schema": {
                            "$ref": "#/definitions/model.Response400Error"
                        }
                    },
                    "401": {
                        "description": "未授權",
                        "schema": {
                            "$ref": "#/definitions/model.Response401Error"
                        }
                    },
                    "500": {
                        "description": "系統錯誤",
                        "schema": {
                            "$ref": "#/definitions/model.Response500Error"
                        }
                    }
                }
            }
        },
        "/api/v1/messages/read-all": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "將登入使用者所有未讀通知標記為已讀。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "全部標記已讀",
                "responses": {
                    "200": {
                        "description": "成功全部標記已讀",
                        "schema": {
                            "$ref": "#/definitions/model.RequestMessage"
                        }
                    },
                    "401": {
                        "description": "未授權",
                        "schema": {
                            "$ref": "#/definitions/model.Response401Error"
                        }
                    },
                    "500": {
                        "description": "系統錯誤",
                        "schema": {
                            "$ref": "#/definitions/model.Response500Error"
                        }
                    }
                }
            }
        },
        "/api/v1/messages/send": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理者傳送社區訊息。系統會先為每位收件者建立 message_info 寄送紀錄，再嘗試透過 FCM 推播；Firebase 未初始化或部分推播失敗時仍會保留寄送紀錄並於回應中標示失敗數量。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "傳送訊息",
                "parameters": [
                    {
                        "description": "訊息資料",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/message_model.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "訊息送出完成",
                        "schema": {
                            "$ref": "#/definitions/message_model.SendMessageResponse"
                        }
                    },
                    "400": {
                        "description": "輸入資料格式錯誤",
                        "schema": {
                            "$ref": "#/definitions/model.Response400Error"
                        }
                    },
                    "401": {
                        "description": "未授權，缺少或無效的 JWT Token",
                        "schema": {
                            "$ref": "#/definitions/model.Response401Error"
                        }
                    },
                    "404": {
                        "description": "找不到符合條件的收件者",
                        "schema": {
                            "$ref": "#/definitions/model.Response404Error"
                        }
                    },
                    "500": {
                        "description": "伺服器內部錯誤",
                        "schema": {
                            "$ref": "#/definitions/model.Response500Error"
                        }
                    }
                }
            }
        },
        "/api/v1/messages/{id}/read": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "將登入使用者的指定通知標記為已讀。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "標記通知為已讀",
                "parameters": [
                    {
                        "type": "string",
                        "description": "通知 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功標記已讀",
                        "schema": {
                            "$ref": "#/definitions/model.RequestMessage"
                        }
                    },
                    "401": {
                        "description": "未授權",
                        "schema": {
                            "$ref": "#/definitions/model.Response401Error"
                        }
                    },
                    "404": {
                        "description": "通知不存在",
                        "schema": {
                            "$ref": "#/definitions/model.Response404Error"
                        }
                    },
                    "500": {
                        "description": "系統錯誤",
                        "schema": {
                            "$ref": "#/definitions/model.Response500Error"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sendmessage": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理者傳送社區訊息。系統會先為每位收件者建立 message_info 寄送紀錄，再嘗試透過 FCM 推播；Firebase 未初始化或部分推播失敗時仍會保留寄送紀錄並於回應中標示失敗數量。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "傳送訊息",
                "parameters": [
                    {
                        "description": "訊息資料",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/message_model.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "訊息送出完成",
                        "schema": {
                            "$ref": "#/definitions/message_model.SendMessageResponse"
                        }
                    },
                    "400": {
                        "description": "輸入資料格式錯誤",
                        "schema": {
                            "$ref": "#/definitions/model.Response400Error"
                        }
                    },
                    "401": {
                        "description": "未授權，缺少或無效的 JWT Token",
                        "schema": {
                            "$ref": "#/definitions/model.Response401Error"
                        }
                    },
                    "404": {
                        "description": "找不到符合條件的收件者",
                        "schema": {
                            "$ref": "#/definitions/model.Response404Error"
                        }
                    },
                    "500": {
                        "description": "伺服器內部錯誤",
                        "schema": {
                            "$ref": "#/definitions/model.Response500Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "message_model.MessageListResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/message_model.MessageSummary"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "message_model.MessageSummary": {
            "type": "object",
            "properties": {
                "create_time": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_read": {
                    "type": "boolean"
                },
                "subtile": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "message_model.SendMessageMetadata": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "announcement"
                },
                "click_action": {
                    "type": "string",
                    "example": "FLUTTER_NOTIFICATION_CLICK"
                },
                "entity_id": {
                    "type": "string",
                    "example": "notice-1"
                },
                "entity_type": {
                    "type": "string",
                    "example": "community"
                }
            }
        },
        "message_model.SendMessageRequest": {
            "type": "object",
            "required": [
                "body",
                "title"
            ],
            "properties": {
                "Userselect": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "user1@example.com",
                        "user2@example.com"
                    ]
                },
                "body": {
                    "type": "string",
                    "example": "您的包裹已抵達管理室，請抽空領取。"
                },
                "deviceToken": {
                    "type": "string",
                    "example": "fcm-device-token"
                },
                "metadata": {
                    "$ref": "#/definitions/message_model.SendMessageMetadata"
                },
                "recipient_emails": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "user1@example.com",
                        "user2@example.com"
                    ]
                },
                "recipient_user_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "user-uuid-1",
                        "user-uuid-2"
                    ]
                },
                "subtitle": {
                    "type": "string",
                    "example": "包裹到達通知"
                },
                "target_type": {
                    "type": "string",
                    "example": "selected_users"
                },
                "title": {
                    "type": "string",
                    "example": "您有新的包裹"
                }
            }
        },
        "message_model.SendMessageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/message_model.SendMessageResponseData"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "message_model.SendMessageResponseData": {
            "type": "object",
            "properties": {
                "created_message_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "failure_count": {
                    "type": "integer"
                },
                "fcm_results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/message_model.SendMessageResult"
                    }
                },
                "message_batch_id": {
                    "type": "string"
                },
                "success_count": {
                    "type": "integer"
                },
                "target_count": {
                    "type": "integer"
                }
            }
        },
        "message_model.SendMessageResult": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "fcm_message_id": {
                    "type": "string"
                },
                "fcm_status": {
                    "type": "string"
                },
                "message_id": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "model.RequestMessage": {
            "type": "object",
            "properties": {
                "Message": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "model.Response400Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Invalid input data"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1001
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Bad Request"
                }
            }
        },
        "model.Response401Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 401
                },
                "error": {
                    "type": "string",
                    "example": "Invalid or missing JWT token"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1002
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Unauthorized"
                }
            }
        },
...
        "model.Response404Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "error": {
                    "type": "string",
                    "example": "Resource not found"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1004
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Not Found"
                }
            }
        },
...
        "model.Response500Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "error": {
                    "type": "string",
                    "example": "Internal server error"
                },
                "internal_code": {
                    "type": "integer",
                    "example": 1000
                },
                "request_id": {
                    "type": "string",
                    "example": "uuid-string"
                },
                "status": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}